+++ jax_galsim/__init__.py
@@ -1,10 +1,16 @@
 # First some basic building blocks that don't usually depend on anything else
 from jax_galsim.position import Position, PositionD, PositionI
+from galsim.angle import radians, hours, degrees, arcmin, arcsec
 
 # GSObject
-from jax_galsim.gsobject import GSObject
 from jax_galsim.gsparams import GSParams
+from jax_galsim.gsobject import GSObject
 from jax_galsim.gaussian import Gaussian
 from jax_galsim.exponential import Exponential
 
+# Derived from GSObjects
 from jax_galsim.sum import Add, Sum
+from jax_galsim.transform import Transformation, Transform
+
+# WCS
+from jax_galsim.wcs import PixelScale, OffsetWCS, JacobianWCS, AffineTransform