+++ CHANGELOG.md
@@ -7,6 +7,9 @@
     * `GSObjects` and `gsparams`
     * `Sum`
     * `Position`, `PositionD` and `PositionI`
+    * `Transformation`
   * Added implementation of simple light profiles:
     * `Gaussian`, `Exponential`
+  * Added implementation of simple WCS:
+    * `PixelScale`, `OffsetWCS`, `JacobianWCS`, `AffineTransform`
   * Added automated suite of tests against reference GalSim